Update UVA 11340_Newspaper.cpp

+++ UVA 11340_Newspaper.cpp
@@ -6,7 +6,7 @@
 
 using namespace std;
 
-typedef pair<char, int>pr;
+typedef pair<char, double>pr;
 
 int main()
 {
